+++ src/sinks/datadog/metrics/encoder.rs
@@ -433,7 +433,7 @@ fn generate_series_metrics(
     let tags = Some(encode_tags(&tags));
     let interval = last_sent
         .map(|then| then.elapsed())
-        .map(|d| d.as_secs().try_into().unwrap_or(i64::MAX));
+        .map(|d| d.as_secs().try_into().unwrap_or(u64::MAX));
 
     let results = match (metric.value(), metric.interval_ms()) {
         (MetricValue::Counter { value }, None) => vec![DatadogSeriesMetric {
@@ -449,8 +449,8 @@ fn generate_series_metrics(
         (MetricValue::Counter { value }, Some(i)) => vec![DatadogSeriesMetric {
             metric: name,
             r#type: DatadogMetricType::Rate,
-            interval: Some(i as i64),
-            points: vec![DatadogPoint(ts, (*value) * 1000 / (i as f64))],
+            interval: Some(i),
+            points: vec![DatadogPoint(ts, (*value) * 1000.0 / (i as f64))],
             tags,
             host,
             source_type_name,
